Added the board score evaluation function for AI.
The function needs to be improved. Lets say you have

*
* x
o o x x
o o o x x o

this configuration. Where x is the AI.

The diagonal is not going to be helpful until o or x places one more disk
in the first column. But this function will give it +100 score. Score should
be assigned after checking if it is possible to win by placing only one disk
in the next move for AI or after a couple of moves.