Board evalution function is now improved. It checks situations like
          1
       *1 2
2 2 2 1 1 1 2

In this configuration, the marked 1 requires 3 more moves to be made to
get the diagonal 4 configuration (of course, the last one being from player 1).

Like

            1
          1 1
       *1 2 2
2 2 2 1 1 1 2

The evaluator now assigns the appropriate score to this state, knowing that
victory is not immediate in the case shown in the beginning.

Same goes for configurations like

*1 1
 2 2 0 0

0 stands for empty position.

Now there are 4 moves before the position marked can lead to victory.

*1 1 1 1
 2 2 2 2

Same goes for left, other diagonal and up directions.